make EVMSmartWallet AA provider agnostic
